requirements.txt
slack-sdk==3.27.1
python-dotenv==1.0.1 